Got Detonation map to display in tabbed pane Signed-off-by:Arie <arieclaassens@gmail.com>

+++ JavaGIS/src/BacterialBomb/Storage.java
@@ -125,7 +125,10 @@ public class Storage {
         for (int i = 0; i < data.length; i++) {
             //inner loop for columns
             for (int j = 0; j < data[i].length; j++) {
-                data[i][j] = Math.random() * 1000;
+                //generate random integer values.
+                //Stoage object in this app uses double, but we're working with INT values and can reduce output
+                //from 1.64MB to 606KB by casting to INT
+                data[i][j] = (int)(Math.random() * 1000);
             }
         }
     }
